Fixing delay controllers
